backend/backend/requirement.txt
asgiref==3.8.1
django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
djangorestframework-simplejwt-token-blacklist==0.1.0
python-decouple==3.8
pyjwt==2.9.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025a
